Update 2021-09-20.scalibility.notify-short-code.md
Added an update to include recommendations from AWS regarding use cases for short codes, and recommended level of documentation of use cases, opt-in/opt-out processes and templates. 

Included in the update the final decision to pause the short code application until such a time as we have a use case that fits the specifics of the short code program that would be acceptable to AWS, CWTA and the carriers.